visual: extract ROP view from vmenus.c to rop.c
